move distr/UpdateParams in independent file
