+++ modules/distribution/handle_block.go
@@ -3,9 +3,6 @@ package distribution
 import (
 	juno "github.com/forbole/juno/v2/types"
 
-	"github.com/forbole/bdjuno/v2/types"
-
-	"github.com/rs/zerolog/log"
 	tmctypes "github.com/tendermint/tendermint/rpc/core/types"
 )
 
@@ -24,25 +21,4 @@ func (m *Module) HandleBlock(
 	}
 
 	return nil
-}
-
-func (m *Module) UpdateParams(height int64) {
-	log.Debug().Str("module", "distribution").Int64("height", height).
-		Msg("updating params")
-
-	params, err := m.source.Params(height)
-	if err != nil {
-		log.Error().Str("module", "distribution").Err(err).
-			Int64("height", height).
-			Msg("error while getting params")
-		return
-	}
-
-	err = m.db.SaveDistributionParams(types.NewDistributionParams(params, height))
-	if err != nil {
-		log.Error().Str("module", "distribution").Err(err).
-			Int64("height", height).
-			Msg("error while saving params")
-		return
-	}
 }
